# backend/viki-openapi.json
{"openapi":"3.1.0","info":{"title":"VIKI API","description":"VIKI social media and creator economy API","version":"1.1.0"},"paths":{"/monetization/gift":{"post":{"tags":["Monetization"],"summary":"Send Gift","operationId":"send_gift_monetization_gift_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GiftRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/monetization/creator/earnings":{"get":{"tags":["Monetization"],"summary":"Creator Earnings","operationId":"creator_earnings_monetization_creator_earnings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/monetization/owner/revenue":{"get":{"tags":["Monetization"],"summary":"Owner Revenue","operationId":"owner_revenue_monetization_owner_revenue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/discover/users":{"get":{"tags":["Discover"],"summary":"Search Users","operationId":"search_users_discover_users_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","default":"","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/discover/follow/{user_id}":{"post":{"tags":["Discover"],"summary":"Follow User","operationId":"follow_user_discover_follow__user_id__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Discover"],"summary":"Unfollow User","operationId":"unfollow_user_discover_follow__user_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profiles/{username}":{"get":{"tags":["Profiles"],"summary":"Get Profile","operationId":"get_profile_profiles__username__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rewards/balance":{"get":{"tags":["Rewards"],"summary":"Reward Balance","operationId":"reward_balance_rewards_balance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/rewards/history":{"get":{"tags":["Rewards"],"summary":"Reward History","operationId":"reward_history_rewards_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/rewards/daily":{"post":{"tags":["Rewards"],"summary":"Daily Reward","operationId":"daily_reward_rewards_daily_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/register":{"post":{"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/me":{"get":{"summary":"Me","operationId":"me_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/users/{username}":{"get":{"summary":"Get Profile","operationId":"get_profile_users__username__get","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{username}/follow":{"post":{"summary":"Follow User","operationId":"follow_user_users__username__follow_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Unfollow User","operationId":"unfollow_user_users__username__follow_delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts":{"post":{"summary":"Create Post","operationId":"create_post_posts_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/posts/{post_id}":{"get":{"summary":"Get Post","operationId":"get_post_posts__post_id__get","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Post","operationId":"delete_post_posts__post_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/feed":{"get":{"summary":"Feed","operationId":"feed_feed_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts/{post_id}/like":{"post":{"summary":"Like Post","operationId":"like_post_posts__post_id__like_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Unlike Post","operationId":"unlike_post_posts__post_id__like_delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts/{post_id}/comments":{"post":{"summary":"Create Comment","operationId":"create_comment_posts__post_id__comments_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Comments","operationId":"get_comments_posts__post_id__comments_get","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/beneficiaries":{"get":{"summary":"List Beneficiaries","operationId":"list_beneficiaries_beneficiaries_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Create Beneficiary","operationId":"create_beneficiary_beneficiaries_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeneficiaryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/beneficiaries/{beneficiary_id}":{"delete":{"summary":"Delete Beneficiary","operationId":"delete_beneficiary_beneficiaries__beneficiary_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"beneficiary_id","in":"path","required":true,"schema":{"type":"integer","title":"Beneficiary Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/wallets/withdraw":{"post":{"summary":"Create Withdrawal","operationId":"create_withdrawal_wallets_withdraw_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/wallets/withdrawals":{"get":{"summary":"List Withdrawals","operationId":"list_withdrawals_wallets_withdrawals_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/currencies":{"get":{"summary":"Currencies","operationId":"currencies_currencies_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/wallets/create":{"post":{"summary":"Create Wallet","operationId":"create_wallet_wallets_create_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"currency","in":"query","required":false,"schema":{"type":"string","default":"USD","title":"Currency"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/wallets":{"get":{"summary":"Wallets","operationId":"wallets_wallets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/wallets/deposit":{"post":{"summary":"Create Deposit","operationId":"create_deposit_wallets_deposit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepositRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/wallets/exchange":{"post":{"summary":"Exchange Wallet","operationId":"exchange_wallet_wallets_exchange_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExchangeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/wallet":{"get":{"summary":"Wallet","operationId":"wallet_wallet_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WalletResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/admin/dev/fund-wallet":{"post":{"summary":"Dev Fund Wallet","description":"DEVELOPMENT ONLY.\n\nAllows an administrator to credit a user's wallet for local\ndevelopment/testing. This is not a real-money deposit system.","operationId":"dev_fund_wallet_admin_dev_fund_wallet_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string","title":"Username"}},{"name":"amount","in":"query","required":false,"schema":{"type":"number","default":10.0,"title":"Amount"}},{"name":"currency","in":"query","required":false,"schema":{"type":"string","default":"USD","title":"Currency"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/wallet/ledger":{"get":{"summary":"Ledger","operationId":"ledger_wallet_ledger_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"BeneficiaryRequest":{"properties":{"beneficiary_type":{"type":"string","maxLength":30,"minLength":3,"title":"Beneficiary Type"},"country":{"type":"string","maxLength":2,"minLength":2,"title":"Country"},"currency":{"type":"string","maxLength":3,"minLength":3,"title":"Currency"},"provider":{"type":"string","maxLength":50,"minLength":2,"title":"Provider"},"account_name":{"type":"string","maxLength":150,"minLength":2,"title":"Account Name"},"account_number":{"type":"string","maxLength":150,"minLength":3,"title":"Account Number"}},"type":"object","required":["beneficiary_type","country","currency","provider","account_name","account_number"],"title":"BeneficiaryRequest"},"CommentCreate":{"properties":{"content":{"type":"string","maxLength":2000,"minLength":1,"title":"Content"}},"type":"object","required":["content"],"title":"CommentCreate"},"DepositRequest":{"properties":{"currency":{"type":"string","maxLength":3,"minLength":3,"title":"Currency"},"amount":{"type":"number","maximum":100000.0,"exclusiveMinimum":0.0,"title":"Amount"},"provider":{"type":"string","maxLength":50,"minLength":2,"title":"Provider"}},"type":"object","required":["currency","amount","provider"],"title":"DepositRequest"},"ExchangeRequest":{"properties":{"from_currency":{"type":"string","maxLength":3,"minLength":3,"title":"From Currency"},"to_currency":{"type":"string","maxLength":3,"minLength":3,"title":"To Currency"},"amount":{"type":"number","exclusiveMinimum":0.0,"title":"Amount"}},"type":"object","required":["from_currency","to_currency","amount"],"title":"ExchangeRequest"},"GiftRequest":{"properties":{"creator_username":{"type":"string","title":"Creator Username"},"amount":{"type":"number","maximum":10000.0,"exclusiveMinimum":0.0,"title":"Amount"},"currency":{"type":"string","maxLength":3,"minLength":3,"title":"Currency","default":"USD"}},"type":"object","required":["creator_username","amount"],"title":"GiftRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"LoginResponse"},"PostCreate":{"properties":{"content":{"type":"string","maxLength":5000,"minLength":1,"title":"Content"},"media_url":{"type":"string","maxLength":1000,"title":"Media Url","default":""},"media_type":{"type":"string","maxLength":20,"title":"Media Type","default":"none"}},"type":"object","required":["content"],"title":"PostCreate"},"RegisterRequest":{"properties":{"username":{"type":"string","maxLength":30,"minLength":3,"title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"},"full_name":{"type":"string","maxLength":100,"minLength":1,"title":"Full Name"},"currency":{"type":"string","maxLength":3,"minLength":3,"title":"Currency","default":"USD"}},"type":"object","required":["username","email","password","full_name"],"title":"RegisterRequest"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"bio":{"type":"string","title":"Bio"},"avatar_url":{"type":"string","title":"Avatar Url"},"is_creator":{"type":"boolean","title":"Is Creator"},"created_at":{"title":"Created At"}},"type":"object","required":["id","username","email","full_name","bio","avatar_url","is_creator","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WalletResponse":{"properties":{"currency":{"type":"string","title":"Currency"},"available":{"type":"number","title":"Available"},"pending":{"type":"number","title":"Pending"}},"type":"object","required":["currency","available","pending"],"title":"WalletResponse"},"WithdrawalRequest":{"properties":{"beneficiary_id":{"type":"integer","exclusiveMinimum":0.0,"title":"Beneficiary Id"},"amount":{"type":"number","maximum":100000.0,"exclusiveMinimum":0.0,"title":"Amount"}},"type":"object","required":["beneficiary_id","amount"],"title":"WithdrawalRequest"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}